リファクタリング: 関数を切り出し
